Fix GitHub Actions CI: Remove cache parameter from Node.js setup
The workflow was failing because actions/setup-node with cache: 'yarn'
was trying to use yarn before corepack was enabled. By removing the cache
parameter, we allow corepack to be enabled first, then yarn can be used
for dependency installation.

This fixes the error: "packageManager": "yarn@4.0.0" but current global
version is 1.22.22 - Corepack must be enabled first.